Remove uthash dependency
By restructuring how we represent IRQs, we remove the need for an
internal hashtable. IRQs are now represented as a 0 indexed per-AFU
number, and are assigned sequentially in the order they are allocated.

This facilitates using a switch statement to handle IRQs.

IRQs can no longer be individually freed, but are still automatically
cleaned when the AFU handle is closed.

As the handle no longer represents the 64 bit ID to pass to the AFU, a
new call ocxl_afu_irq_get_id() has been introduced to retrieve the ID
from the IRQ.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>