ocxl_memcpy: Add memory barrier
It's theoretically possible for the core to reorder some instructions
and start loading the source and destination buffers for the
comparison before the work element status is set to complete. The
hardware cannot guess the data dependency if we don't spell it out.

It seems a bit far-fetched here, as it would require some amount of
branch predictions. But it was apparently hit on a slightly simpler
version of this test used by HTX.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>