Makefile: Fix header file detection with latest versions of make
Gnu make 4.3 introduced a backward-incompatible change in its handling
of the escape character before # when calling a function. And it is
used when writing some test code to check the latest version of the
ocxl header file (#include <misc/ocxl.h>).
This patch reworks how the test code is passed to the compiler to
avoid complications and be compatible with all version of make.

Signed-off-by: Frederic Barrat <fbarrat@linux.ibm.com>